Informative error msg for dataloading an empty dir (#109)
The `ScalableShardDataset` assumes that at least one of its sub-loaders owns at least one document, and when this is not the case, it errors out via the `torch.multinomial()` call. The resulting error message is not particularly clear or useful, so add a simple assert message to catch this case and inform the user that none of the workers were able to find any documents to load.

Signed-off-by: Joshua Rosenkranz <jmrosenk@us.ibm.com>